backend_only/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
aiofiles==23.2.1
python-docx==0.8.11
markdown==3.5.1
reportlab==4.0.7
aiosqlite==0.17.0
gunicorn
aiomysql==0.2.0
PyMySQL==1.1.0
cryptography==41.0.7